Changes to routes, moved into api folder, added index files along with reaction route and modified server.js file
