+++ routes/api/thoughtRoutes.js
@@ -1,7 +1,7 @@
 // Dependencies for MongoDB
 const express = require('express');
 const router = express.Router();
-const thoughtController = require('../controllers/thoughtController');
+const thoughtController = require('../../controllers/thoughtController');
 
 // GET all thoughts
 router.get('/', thoughtController.getAllThoughts);
