+++ routes/api/userRoutes.js
@@ -1,7 +1,7 @@
 // Dependencies for MongoDB
 const express = require('express');
 const router = express.Router();
-const userController = require('../controllers/userController');
+const userController = require('../../controllers/userController');
 
 // GET all users
 router.get('/', userController.getAllUsers);
